requirements.txt
discord.py>=2.3.2
supabase>=1.1.1
python-dotenv>=1.0.0
asyncpg>=0.29.0
aiohttp>=3.9.0
dataclasses>=0.6  # لـ Python 3.6
Pillow>=10.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
hijri-converter>=2.3.1